+++ src/components/custom-field-input-single-choice/custom-field-input-single-choice.jsx
@@ -1,5 +1,4 @@
 import React, {
-  useEffect,
   useRef,
   useState,
 } from 'react';
@@ -12,9 +11,9 @@ import iconCaretDownDisabled from '../../svgs/icon-caret-down-disabled.svg';
 import styles from './custom-field-input-single-choice.css';
 import Listbox from '../listbox/listbox.jsx';
 import ListOption from '../list-option/list-option.jsx';
+import useDropdownClose from '../../hooks/use-dropdown-close.js';
 
 export default function CustomFieldInputSingleChoice(props) {
-  const [didMount, setDidMount] = useState(false);
   const [showOptions, setShowOptions] = useState(false);
   const [value, setValue] = useState(props.value);
   const [searchValue, setSearchValue] = useState(undefined);
@@ -26,6 +25,15 @@ export default function CustomFieldInputSingleChoice(props) {
     name={props.readOnly ? iconCaretDownDisabled.id : iconCaretDown.id}
     fill="skip"
   />);
+
+  const defaultValue = value ? value.label : '';
+
+  const wrapperRef = useRef(null);
+  const handleDropdownClose = () => {
+    setShowOptions(false);
+    setSearchValue(defaultValue);
+  };
+  useDropdownClose(wrapperRef, showOptions, handleDropdownClose);
 
   const clear = () => {
     setValue(undefined);
@@ -98,6 +106,7 @@ export default function CustomFieldInputSingleChoice(props) {
     setValue(selectedValue);
     setSearchValue(selectedValue.label);
     setShowOptions(false);
+    inputRef.current.focus();
   }
 
   function onSearchChange(event) {
@@ -113,17 +122,8 @@ export default function CustomFieldInputSingleChoice(props) {
     setShowOptions(true);
   }
 
-  useEffect(() => {
-    setDidMount(true);
-  }, []);
-
-  useEffect(() => {
-    if (!didMount) return;
-    if (!showOptions) inputRef.current.focus();
-  }, [showOptions]);
-
   return (
-    <div className={styles.container}>
+    <div ref={wrapperRef} className={styles.container}>
       <CustomFieldInputText
         icon={caretIcon}
         clear={clearIcon()}
@@ -138,7 +138,7 @@ export default function CustomFieldInputSingleChoice(props) {
         required={props.required}
         error={props.error}
         helpText={props.helpText}
-        value={searchValue || (value ? value.label : '')}
+        value={searchValue || defaultValue}
       />
       { showOptions && (
         <Listbox